Roll custom printfloat that seem to be working

+++ lib/printfloat/printfloat.h
@@ -2,6 +2,6 @@
 #include <stddef.h>
 #include <stdint.h>
 
-size_t snprintfloat(char* dst, size_t sz, float f, uint8_t fraq=3, uint8_t len=0);
+int snprintfloat(char* dst, size_t sz, float f, uint8_t fraq=3, uint8_t len=0);
 
 double _atod(const char* std);
